# UC-US-BA-APIDESIGN/openapi.yaml
openapi: 3.0.3
info:
  title: Booking API — Hệ thống booking massage りらくる
  description: |
    BE cung cấp API cho FE web (và AI chatbot ở giai đoạn 2).
    Mã lỗi map với Business Rules — xem api-design.md.
  version: 0.1.0
servers:
  - url: /api/v1
...
paths:
...
  /shops/{shopId}/services:
    get:
      tags: [booking-flow]
      summary: Dịch vụ khả dụng theo ngày (bước 4–5, case A1)
      parameters:
        - { name: shopId, in: path, required: true, schema: { type: integer } }
        - { name: date, in: query, required: true, schema: { type: string, format: date } }
        - { name: party_size, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 3 } }
      responses:
        "200":
          description: Rỗng + reason=SHOP_CLOSED nếu ngày đó không phục vụ (A1)
          content:
            application/json:
              schema:
                type: object
                properties:
                  courses:
                    type: array
                    items: { $ref: "#/components/schemas/Course" }
                  addons:
                    type: array
                    items: { $ref: "#/components/schemas/Addon" }
                  reason:
                    type: string
                    nullable: true
                    enum: [SHOP_CLOSED, null]
  /shops/{shopId}/slots:
    get:
      tags: [booking-flow]
      summary: Tra slot khả dụng (bước 6 — UC-04, BR-07). Chỉ mang tính hiển thị, BE re-check lúc tạo (BR-08)
      parameters:
        - { name: shopId, in: path, required: true, schema: { type: integer } }
        - { name: date, in: query, required: true, schema: { type: string, format: date } }
        - { name: party_size, in: query, required: true, schema: { type: integer, minimum: 1, maximum: 3 } }
        - { name: course_id, in: query, required: true, schema: { type: integer } }
        - { name: addon_ids, in: query, required: false, style: form, explode: false,
            schema: { type: array, items: { type: integer } } }
        - { name: therapist_id, in: query, required: false, schema: { type: integer },
            description: Chỉ booking 1 người (BR-04); loại trừ với therapist_gender }
        - { name: therapist_gender, in: query, required: false,
            schema: { $ref: "#/components/schemas/Gender" } }
      responses:
        "200":
          description: Rỗng = hết slot (case A2)
          content:
            application/json:
              schema:
                type: object
                properties:
                  slots:
                    type: array
                    items: { type: string, example: "14:00" }
  /shops/{shopId}/therapists:
    get:
      tags: [booking-flow]
      summary: Danh sách therapist cho dropdown chỉ định theo tên (màn 3, chỉ booking 1 người — BR-04)
      parameters:
        - { name: shopId, in: path, required: true, schema: { type: integer } }
        - { name: date, in: query, required: false, schema: { type: string, format: date },
            description: Lọc người có ca ngày đó }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: integer }
                    name: { type: string }
                    gender: { $ref: "#/components/schemas/Gender" }
  /customers/lookup:
    post:
      tags: [booking-flow]
      summary: Nhận dạng khách qua SĐT (bước 9 — UC-05, UC-06). Dùng POST để SĐT không lộ trên URL/log
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string, example: "09012345678" }
      responses:
        "200":
          description: Thông tin thành viên (rank chỉ hiển thị — BR-20)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerInfo" }
        "403":
          description: SĐT trong NG list (BR-06), details.reason có lý do cấm (BR-20)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /bookings:
    post:
      tags: [booking-flow]
      summary: Tạo booking (bước 11 — UC-01). Transaction + lock slot (BR-08, BR-12)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookingCreateRequest" }
      responses:
        "201":
          description: Tạo thành công — mã đặt chỗ gửi kèm email qua SES (BR-15)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BookingCreated" }
        "400":
          description: PARTY_SIZE_EXCEEDED (BR-14) / ADDON_WITHOUT_COURSE (BR-01) / THERAPIST_NOT_ALLOWED (BR-04)
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "403":
          description: PHONE_BLOCKED (BR-06)
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "409":
          description: SLOT_CONFLICT (BR-08) — details.suggested_slots gợi ý 3 giờ gần nhất (case A6)
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "422":
          description: INVALID_COMBO (BR-09) / THERAPIST_OFF_SHIFT (BR-05)
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "500":
          description: INTERNAL_ERROR (BR-12) — khách thử lại
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    editToken:
      type: apiKey
      in: header
      name: X-Edit-Token
      description: JWT ngắn hạn {booking_id, exp +120s} — BE cấp lúc tạo booking (BR-17)
  schemas:
    Gender:
      type: string
      enum: [male, female]
    BookingStatus:
      type: string
      enum: [pending, confirmed, cancelled, completed, no_show]
    Shop:
      type: object
      properties:
        id: { type: integer }
        shop_code: { type: string, example: "S001" }
        name: { type: string }
        address: { type: string }
        phone: { type: string }
    Course:
      type: object
      properties:
        id: { type: integer }
        name: { type: string, example: "もみほぐし" }
        duration_min: { type: integer, example: 60, description: Bội số 15 phút (BR-02) }
        price: { type: integer, example: 3980, description: JPY }
    Addon:
      allOf:
        - $ref: "#/components/schemas/Course"
        - type: object
          properties:
            restricted_course_ids:
              type: array
              items: { type: integer }
              description: Các course không được ghép (từ combo_restriction — BR-09, FE chặn sớm)
...
    CustomerInfo:
      type: object
      properties:
        member_type: { type: string, enum: [member, guest] }
        rank: { type: string, nullable: true, description: Chỉ để hiển thị (BR-20) }
        visit_count: { type: integer }
    ReservationInput:
      type: object
      properties:
        addon_ids:
          type: array
          items: { type: integer }
          description: Add-on riêng từng người (BR-10)
    BookingCreateRequest:
      type: object
      required: [shop_id, date, start_time, party_size, phone, email, course_id, reservations]
      properties:
        shop_id: { type: integer }
        date: { type: string, format: date }
        start_time: { type: string, example: "14:00" }
        party_size:
          type: integer
          minimum: 1
          maximum: 3
          description: Tối đa 3 (BR-14), phải khớp độ dài reservations[]
        phone: { type: string, description: Tra thành viên / NG list (BR-06) }
        email: { type: string, format: email, description: Xác thực + nhận mã (BR-15) }
        course_id: { type: integer, description: Cả nhóm cùng course (BR-10) }
        reservations:
          type: array
          minItems: 1
          maxItems: 3
          items: { $ref: "#/components/schemas/ReservationInput" }
        therapist_id:
          type: integer
          nullable: true
          description: Chỉ booking 1 người (BR-04); loại trừ với therapist_gender
        therapist_gender:
          allOf: [{ $ref: "#/components/schemas/Gender" }]
          nullable: true
...
    Booking:
      type: object
      properties:
        booking_code: { type: string, example: "20260720-S001-A1B2" }
        status: { $ref: "#/components/schemas/BookingStatus" }
        shop: { $ref: "#/components/schemas/Shop" }
        date: { type: string, format: date }
        start_time: { type: string }
        party_size: { type: integer }
        course: { $ref: "#/components/schemas/Course" }
        reservations:
          type: array
          items:
            type: object
            properties:
              guest_no: { type: integer }
              addons:
                type: array
                items: { $ref: "#/components/schemas/Course" }
        therapist_name: { type: string, nullable: true }
        can_modify:
          type: boolean
          description: BE tính sẵn theo deadline 1h (BR-16) để FE ẩn/hiện nút
    BookingCreated:
      allOf:
        - $ref: "#/components/schemas/Booking"
        - type: object
          properties:
            edit_token:
              type: string
              description: Token sửa nhanh, hiệu lực 2 phút (BR-17)
            edit_token_expires_in:
              type: integer
              example: 120
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - PARTY_SIZE_EXCEEDED
                - ADDON_WITHOUT_COURSE
                - THERAPIST_NOT_ALLOWED
                - PHONE_BLOCKED
                - BOOKING_NOT_FOUND
                - SLOT_CONFLICT
                - INVALID_COMBO
                - THERAPIST_OFF_SHIFT
                - MODIFY_DEADLINE_PASSED
                - SHOP_CHANGE_NOT_ALLOWED
                - EDIT_TOKEN_EXPIRED
                - RATE_LIMITED
                - INTERNAL_ERROR
            message: { type: string }
            details:
              type: object
              nullable: true
              description: "VD: SLOT_CONFLICT kèm suggested_slots[], PHONE_BLOCKED kèm reason"